pyproject.toml
[project]
name = "hackathon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "requests>=2.31.0",
    "django-axes>=8.0.0",
    "django-otp>=1.5.0",
    "qrcode>=7.4.0",  # For TOTP QR codes
    "python-decouple>=3.8",
    "redis>=5.0.0",
    "bleach>=6.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "ruff>=0.14.6",
]
